fix: resolve all mypy strict errors blocking CI on phase1-3 branch
- graph.py: wrap has_node/has_edge/is_dag returns in bool() to fix no-any-return
- propagation_index.py: remove unused type: ignore[name-defined] on build_propagation_index
- fast_kernel.py: drop np=None assignment (np unused at runtime) to fix unused-ignore;
  assign get_propagation_index() result to typed variable to fix no-any-return
- controller.py: fix _visible_context cache param type from dict[float,str] to
  dict[tuple[float,str],str] to match actual (rounded_ratio, node_id) key
- session_store.py: add type: ignore[arg-type] for dict[str,object].get() passed to float()
- metrics.py: annotate build_what_if_k_curve model param as Any (no-untyped-def)
- risk_predictors.py: annotate min_k_for_quality_floor simulate param as Any
- evidence_harness.py: wrap _ci_half_width return in float() to fix no-any-return from pstdev

+++ src/agentprop/evaluation/evidence_harness.py
@@ -143,4 +143,4 @@ def summaries_to_dict(summaries: list[EvidenceSummary]) -> list[dict[str, Any]]:
 def _ci_half_width(values: list[float], z: float = 1.96) -> float:
     if len(values) < 2:
         return 0.0
-    return z * pstdev(values) / (len(values) ** 0.5)
+    return float(z * pstdev(values) / (len(values) ** 0.5))
